refactor: Moved app init in a separate file
